Some restructuring, and PEP8
